Change updates endpoint to use most recent updates for each vehicle
Previously, the endpoint looped through each vehicle's last 20 updates
to determine if it has been stationary for more than a few minutes, and
then only returned the latest update for each non-stationary vehicle.

This is a lot of work to do on each request, but instead we can just
look at the most recent update for each vehicle and return any that is
from the last five minutes.

This works because as of 91d7de4 we only store updates if they are new,
instead of storing a new update every update interval regardless of
whether there's a change. iTrak appears to only return new updates if
the vehicle is not stationary.

Also changed the vehicle `active` flag to be `enabled`. This better
reflects the intent for it to be admin-controlled instead of determined
by the shuttle's own activity.