Create a dark theme for the app.
This theme is CSS-based. It is activated by adding an attribute to the <body> tag: `data-theme="dark"`. The definitions for this theme are in
vars.css. All color references in the app are replaced with CSS variables that target the current theme. Bulma style variables are also
overridden to comply with the theme.