Dark Theme (#276)
* Add a setting for dark theme (setting ONLY, no actual theme).

* Create a dark theme for the app.

This theme is CSS-based. It is activated by adding an attribute to the <body> tag: `data-theme="dark"`. The definitions for this theme are in
vars.css. All color references in the app are replaced with CSS variables that target the current theme. Bulma style variables are also
overridden to comply with the theme.

* Connect the dark theme settings with the dark theme CSS

* Add dark map tiles for the dark theme